LF-5478 Organize files under a feature component folder
