+++ packages/webapp/src/components/Insights/Survey/SurveyBandLegend/styles.module.scss
