+++ packages/webapp/src/components/Insights/Survey/SurveyModuleCard/Gauge/styles.module.scss
