+++ packages/webapp/src/components/Insights/Survey/SurveyModuleCard/styles.module.scss
