+++ packages/webapp/src/components/Insights/Survey/surveyScoreBand.ts
